Run hbs on server tests
