+++ packages/server/src/jsRunner/tests/jsRunner.spec.ts
@@ -1,21 +1,35 @@
+jest.mock("@budibase/handlebars-helpers/lib/math", () => {
+  const actual = jest.requireActual("@budibase/handlebars-helpers/lib/math")
+
+  return {
+    ...actual,
+    random: () => 10,
+  }
+})
+jest.mock("@budibase/handlebars-helpers/lib/uuid", () => {
+  const actual = jest.requireActual("@budibase/handlebars-helpers/lib/uuid")
+
+  return {
+    ...actual,
+    uuid: () => "f34ebc66-93bd-4f7c-b79b-92b5569138bc",
+  }
+})
+
 import { processStringSync, encodeJSBinding } from "@budibase/string-templates"
-import TestConfiguration from "../../tests/utilities/TestConfiguration"
+const { runJsHelpersTests } = require("@budibase/string-templates/test/utils")
+
+import tk from "timekeeper"
+tk.freeze("2021-01-21T12:00:00")
 
 describe("jsRunner", () => {
-  const config = new TestConfiguration()
-
-  beforeAll(async () => {
-    await config.init()
-  })
-
   const processJS = (js: string, context?: object) => {
-    return config.doInContext(config.getAppId(), async () =>
-      processStringSync(encodeJSBinding(js), context || {})
-    )
+    return processStringSync(encodeJSBinding(js), context || {})
   }
 
-  it("it can run a basic javascript", async () => {
-    const output = await processJS(`return 1 + 2`)
+  it("it can run a basic javascript", () => {
+    const output = processJS(`return 1 + 2`)
     expect(output).toBe(3)
   })
+
+  runJsHelpersTests()
 })
